Documenta extensoes da camera e corrige indices
